Move stepper_indirection to sub-folder
