+++ Marlin/src/gcode/feature/L6470/M916-918.cpp
@@ -25,7 +25,7 @@
 #if HAS_DRIVER(L6470)
 
 #include "../../gcode.h"
-#include "../../../module/stepper_indirection.h"
+#include "../../../module/stepper/indirection.h"
 #include "../../../module/planner.h"
 #include "../../../libs/L6470/L6470_Marlin.h"
