+++ Marlin/src/lcd/menu/menu_tmc.cpp
@@ -29,7 +29,7 @@
 #if HAS_TRINAMIC && HAS_LCD_MENU
 
 #include "menu.h"
-#include "../../module/stepper_indirection.h"
+#include "../../module/stepper/indirection.h"
 #include "../../feature/tmc_util.h"
 
 #define TMC_EDIT_STORED_I_RMS(ST,MSG) MENU_ITEM_EDIT_CALLBACK(uint16_4, MSG, &stepper##ST.val_mA, 100, 3000, refresh_stepper_current_##ST)
